+++ contracts/LockGTON.sol
@@ -7,7 +7,6 @@ import "./interfaces/ILockGTON.sol";
 /// @author Artemij Artamonov - <array.clean@gmail.com>
 /// @author Anton Davydov - <fetsorn@gmail.com>
 contract LockGTON is ILockGTON {
-
     /// @inheritdoc ILockGTON
     address public override owner;
 
@@ -41,12 +40,8 @@ contract LockGTON is ILockGTON {
     }
 
     /// @inheritdoc ILockGTON
-    function migrate(address newLock)
-        external
-        override
-        isOwner
-    {
-        uint amount = governanceToken.balanceOf(address(this));
+    function migrate(address newLock) external override isOwner {
+        uint256 amount = governanceToken.balanceOf(address(this));
         governanceToken.transfer(newLock, amount);
         emit Migrate(newLock, amount);
     }
